% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.bettertrace.R
\docType{package}
\name{bettertrace}
\alias{bettertrace}
\alias{bettertrace-package}
\title{A better stack trace for R.}
\description{
When this package is loaded, the option
\code{options(error = bettertrace::stacktrace)} will be set, providing
a better and more visually helpful stack trace.
}

