% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stacktrace.R
\name{stacktrace}
\alias{stacktrace}
\title{A traceback function to be used with error handling.}
\usage{
stacktrace()
}
\description{
To use, call \options{error = stacktrace}, which will be called
by default when this package is loaded.
}

