% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacktrace.R
\name{stacktrace}
\alias{stacktrace}
\title{A traceback function to be used with error handling.}
\usage{
stacktrace()
}
\description{
To use, call \code{options(error = stacktrace)}, which will be called
by default when this package is loaded.
}
\examples{
\dontrun{
  # /Users/you/tmp/foo.R
  foo <- function() { bar() + 1 }
  bar <- function() { baz() + 2 }

  # R console
  source("~/tmp/foo.R")
  foo()
  # Error in bar() : could not find function "baz"
  # In global environment: foo()
  # In /Users/robertk/tmp/foo.R:1: bar() + 1
  # In /Users/robertk/tmp/foo.R:2: baz() + 2
  #
  # could not find function "baz"
}
}

